BUG: errors in fast_scalar_power are not propagated.
If one raises an ndarray or subclass to particular powers,
fast_scalar_power is used in number.c -- this defers to another
function (e.g., sqrt for 0.5), but does not check whether an error
occurred. In consequece, if a subclass raises an error, and returns
NULL, that gets interpreted as meaning the fast operation is not
possible, and power is tried in turn, likely raising another
exception.  This commit fixes fast_scalar_power such that it
returns success any time an operation is attempted.