f2py: allow Fortran arrays of dimension 0
Up until now, f2py throw an error when arrays were declared which had
dimension of length 0. This, however, is a perfectly legal case, and in
fact occurs frequently in the context of linear algrebra. This bug was
discovered, for example, in an interface that does matrix
tridiagonalization. If the matrix is 1x1, the super- and subdiagonal are
of length 0.

Note that negative dimensions continue to produce errors although
Fortran also allows this (it always allocates to size max(0, n)).