BUG: Functions in ma should not add Notes to existing docstrings.
Currently, the `doc_note` function in `numpy.ma.core.py` is called to
add a "Notes" section to the existing docstrings of the unmasked
versions of some functions. This may result in duplicate "Notes"
sections, or incorrect placement, which are errors in current numpydoc.
This PR disables the `doc_note` function until we can decide on the
proper way to deal with the problem.