BUG: fix complex casting error in cov with aweights
When using cov with a complex input and with aweights specified, cov
will error as a result of trying to cast a complex value into a float64.
This comes about since average is used to calculate the sum of the
weights from aweights.  average returns the sum of weights as the same
type as its result, not the weights type.  For a complex input m, and
any type for aweights, this would result in a complex value for fact.
It appears the primary purpose of np.float64(fact) is to provide a NaN
value from the divide when fact is an integer zero.  This has been
replaced by using numpy.divide to replicate the same behavior, but to
also handle complex types.