MAINT: Cleanup pavement.py
Remove long unused options from pavement.py. There is actually little
left, and what is there could easily be turned into a python script at
this point, but this clarifies what is currently needed.

- Remove Windows installers code
- Remove OS X dmg installers code
- Remove Documentation generation code
- Remove Changelog generation
- Remove unused variables and make pylint happy

Also fix a Python 3 bug, the files to be hashed need to be
opened as binary files.

[ci skip]