DOC: Change 'num' to 'np'
Stumbled upon this *NumPy for Matlab users* guide. Even though it says *Note: this is just an example, not a statement of "best practices"* I think it would still be better to use `np` instead of `num`, as this is the most common abbreviation you'll find in examples etc.

Just close the PR if you disagree.