BUG/ENH: Make floor, ceil, and trunc call the matching special methods
Previously `np.ceil` would call `o.ceil()` on each element of an object array.

This is inconsistent with the builtin python way of handling this, calling `o.__ceil__()`.

This changes these three functions to use the corresponding functions in the `math` module, which do the special method lookup.

As a result, they now work on arrays of `Fraction` and `Decimal` objects.